fix: Fixed issue with defining User after LogedInUser
